looking/requirements.txt
torch
Pillow
opencv-python
scikit-learn
openpifpaf
numpy
scikit-image